V1.2

+++ logserver/runserver.py
@@ -5,6 +5,11 @@ import pandas as pd
 conn = sqlite3.connect('users.db')
 c = conn.cursor()
 
+@hug.directive()
+def cors(support='*', response=None, **kwargs):
+    '''Returns passed in parameter multiplied by itself'''
+    response and response.set_header('Access-Control-Allow-Origin', support)
+
 @hug.get(examples='username=bgjerstad&compname=011acboe&stat=on&time=2016-10-20_0229 PM')
 @hug.local()
 def log_this(username: hug.types.text, compname: hug.types.text,stat: hug.types.text,time: hug.types.text,  hug_timer=3):
@@ -20,7 +25,7 @@ def log_this(username: hug.types.text, compname: hug.types.text,stat: hug.types.
 
 @hug.get(examples='username=bgjerstad&compname=011acboe')
 @hug.local()
-def get_log(username: hug.types.text, compname: hug.types.text,hug_timer=3):
+def get_log(hug_cors,username: hug.types.text, compname: hug.types.text,hug_timer=3):
 	data2 = (username,compname)
 	print(data2)
 	c = conn.cursor()
@@ -45,7 +50,7 @@ def get_log(username: hug.types.text, compname: hug.types.text,hug_timer=3):
 
 @hug.get(examples='')
 @hug.local()
-def get_dup(hug_timer=3):
+def get_dup(hug_cors,hug_timer=3):
 	logs = {}
 	dbkeys = ['compname','time','stat']
 	#exclustion list
@@ -59,7 +64,6 @@ def get_dup(hug_timer=3):
 		c.execute("SELECT DISTINCT username FROM exclude WHERE 1")
 		exlist = c.fetchall()
 		exlist = [x[0] for x in exlist]
-	print('exlist',exlist)
 	#get a list of all of the users
 	userlist = []
 	try:
@@ -78,37 +82,65 @@ def get_dup(hug_timer=3):
 		logcnt = len(dbout)	
 		if(logcnt > 1):
 			last = ''
+			lastcomp = ''
 			for row in dbout:
-				if(row[2]==last):
+				if(row[2]==last and not(row[0]==lastcomp)):
 					#print(user,row)
 					thisis = user+row[1]
 					logs[thisis] = dict(zip(dbkeys,row))
 				last = row[2]
+				lastcomp = row[0]
 	return logs
 
-@hug.get(examples='action=clear')
+@hug.get(examples='action=clearlog')
 @hug.local()
-def db(action: hug.types.text, hug_timer=3):
-	#drop the table then recreate it.
-	c.execute("DROP TABLE users")
-	c.execute('''Create Table users (username text,compname text,stat text,time text)''')
-	conn.commit()
+def db(hug_cors,action: hug.types.text, hug_timer=3):
+	if (action == 'clearlog'):
+		#drop the table then recreate it.
+		c.execute("DROP TABLE users")
+		c.execute('''Create Table users (username text,compname text,stat text,time text)''')
+		conn.commit()
+	if (action == 'clearex'):
+		#drop the table then recreate it.
+		c.execute("DROP TABLE exclude")
+		c.execute('''Create Table exclude (username text)''')
+		conn.commit()
+	if (action == 'clearall'):
+		#drop the table then recreate it.
+		c.execute("DROP TABLE users")
+		c.execute("DROP TABLE exclude")
+		c.execute('''Create Table users (username text,compname text,stat text,time text)''')
+		c.execute('''Create Table exclude (username text)''')
+		conn.commit()
 	return 1
 
-@hug.get(examples='username=bgjerstad&action=add&lvl=dup')
+@hug.get(examples='username=bgjerstad&action=add')
 @hug.local()
-def ex_this(username: hug.types.text, action: hug.types.text, lvl: hug.types.text,hug_timer=3):
-	data = {'username':'{0}'.format(username),'action':'{0}'.format(action),'lvl':'{0}'.format(lvl)}
-	data2 = (username, lvl)
+def ex_this(hug_cors,username: hug.types.text, action: hug.types.text, hug_timer=3):
+	data = {'username':'{0}'.format(username),'action':'{0}'.format(action)}
+	data2 = (username)
+	logs = {}
+	dbkeys = ['username']
 	if (action == 'add'):
 		try:
-			c.execute("INSERT INTO exclude VALUES "+str(data2))
+			print("INSERT INTO exclude VALUES "+str(data2))
+			c.execute("INSERT INTO exclude VALUES ('"+str(data2)+"')")
 		except sqlite3.OperationalError:
 			makedb()
 			c.execute("INSERT INTO exclude VALUES "+str(data2))
-	data2 = (username, lvl)
+	data2 = (username)
 	if (action == 'remove'):
 		c.execute("DELETE FROM exclude WHERE username='"+str(data2[0])+"'")
+	if (action == 'list'):
+			exlist = []
+			try:
+				dbout = c.execute("SELECT DISTINCT username FROM exclude WHERE 1")
+			except sqlite3.OperationalError:
+				makedb()
+				dbout = c.execute("SELECT DISTINCT username FROM exclude WHERE 1")
+			for idx,row in enumerate(dbout):
+				logs[idx] = dict(zip(dbkeys,row))
+			data = logs
 	conn.commit()
 	return data
 
@@ -116,6 +148,6 @@ def makedb():
 	conn = sqlite3.connect('users.db')
 	c = conn.cursor()
 	c.execute('''Create Table users (username text,compname text,stat text,time text)''')
-	c.execute('''Create Table exclude (username text, lvl text)''')
+	c.execute('''Create Table exclude (username text)''')
 	conn.commit()
